fix code style

+++ modules/redirect.sh
@@ -19,9 +19,9 @@ function redirect {
         ;;
     remove)
         checkargn $# 2
-        for i in $(ls /etc/dnsmasq.d/ --ignore="README")
+        for i in /etc/dnsmasq.d/* 
         do
-          if [ "$i" == "$2" ]; then
+          if [ "$i" == "$2" ] && [ "$2" != "README" ]; then
             rm -rf /etc/dnsmasq.d/$2
             systemctl restart dnsmasq.service
             echo "$2 removed."
